extract mapping logic
