+++ lib/primer/view_components/linters/argument_mappers/conversion_error.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+module Primer
+  module ViewComponents
+    module Linters
+      module ArgumentMappers
+        # Error when converting atguments.
+        class ConversionError < StandardError; end
+      end
+    end
+  end
+end
